+++ breeze-core/src/main/kotlin/component/CaseFormat.kt
@@ -6,11 +6,11 @@ package icu.windea.breezeframework.core.component
 import icu.windea.breezeframework.core.extension.*
 
 /**
- * 字母格式。
+ * 大小写格式
  *
- * 字母格式用于表示单词组的显示格式，基于某种大小写和单词的分割方式。
+ * 大小写格式用于表示单词组的显示格式，基于字母的大小写和单词的分割方式。
  */
-interface LetterCase:Component {
+interface CaseFormat : Component {
 	/**
 	 * 判断指定的字符串是否匹配指定的单词格式。
 	 */
@@ -41,23 +41,23 @@ interface LetterCase:Component {
 	 */
 	fun joinToString(value: Sequence<String>): String
 
-	companion object Registry: AbstractComponentRegistry<LetterCase>(){
+	companion object Registry : AbstractComponentRegistry<CaseFormat>() {
 		override fun registerDefault() {
-			register(LowerCase)
-			register(UpperCase)
+			register(LowerCaseFormat)
+			register(UpperCaseFormat)
 			register(Capitalized)
-			register(LowerCaseWords)
-			register(UpperCaseWords)
+			register(LowerCaseFormatWords)
+			register(UpperCaseFormatWords)
 			register(FirstWordCapitalized)
 			register(CapitalizedWords)
 			register(Words)
-			register(CamelCase)
-			register(PascalCase)
-			register(SnakeCase)
-			register(ScreamingSnakeCase)
+			register(CamelCaseFormat)
+			register(PascalCaseFormat)
+			register(SnakeCaseFormat)
+			register(ScreamingSnakeCaseFormat)
 			register(UnderscoreWords)
-			register(KebabCase)
-			register(KebabUpperCase)
+			register(KebabCaseFormat)
+			register(KebabUpperCaseFormat)
 			register(HyphenWords)
 
 			register(ReferencePath)
@@ -69,7 +69,7 @@ interface LetterCase:Component {
 		 * 推断单词格式。
 		 */
 		@JvmStatic
-		fun infer(value: String): LetterCase? {
+		fun infer(value: String): CaseFormat? {
 			for(letterCase in components) {
 				if(letterCase.matches(value)) return letterCase
 			}
@@ -83,7 +83,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`lowercase`
 	 */
-	object LowerCase : LetterCase {
+	object LowerCaseFormat : CaseFormat {
 		private val regex = """[a-z]+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = listOf(value)
@@ -98,7 +98,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`UPPERCASE`
 	 */
-	object UpperCase : LetterCase {
+	object UpperCaseFormat : CaseFormat {
 		private val regex = """[A-Z]+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = listOf(value)
@@ -113,7 +113,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`Capitalized`
 	 */
-	object Capitalized : LetterCase {
+	object Capitalized : CaseFormat {
 		private val regex = """[A-Z][a-z]+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = listOf(value)
@@ -128,7 +128,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`lowercase words`
 	 */
-	object LowerCaseWords : LetterCase {
+	object LowerCaseFormatWords : CaseFormat {
 		private val regex = """[a-z']+(?:\s+[a-z']+)+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.split(' ').filterNotEmpty()
@@ -143,7 +143,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`UPPERCASE WORDS`
 	 */
-	object UpperCaseWords : LetterCase {
+	object UpperCaseFormatWords : CaseFormat {
 		private val regex = """[A-Z']+(?:\s+[A-Z']+)+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.split(' ').filterNotEmpty()
@@ -158,7 +158,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`Uppercase words`
 	 */
-	object FirstWordCapitalized : LetterCase {
+	object FirstWordCapitalized : CaseFormat {
 		private val regex = """[A-Z][a-z']*(?:\s+[a-z']+)+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.split(' ').filterNotEmpty()
@@ -173,7 +173,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`Uppercase Words`
 	 */
-	object CapitalizedWords : LetterCase {
+	object CapitalizedWords : CaseFormat {
 		private val regex = """[A-Z][a-z']*(?:\s+[a-z']+)+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.split(' ').filterNotEmpty()
@@ -188,7 +188,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`Words words Words`
 	 */
-	object Words : LetterCase {
+	object Words : CaseFormat {
 		private val regex = """[a-zA-Z']+(?:\s+[a-zA-Z']+)+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.split(' ').filterNotEmpty()
@@ -203,22 +203,27 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`camelCase`
 	 */
-	object CamelCase : LetterCase {
+	object CamelCaseFormat : CaseFormat {
 		private val regex = """\$?[a-z]+(?:\$?[A-Z][a-z]+\$?|\$?[A-Z]+\$?|\d+)+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.splitWords().split(' ')
 		override fun splitToSequence(value: String) = value.splitWords().splitToSequence(' ')
-		override fun joinToString(value: Array<String>) = value.joinToString("") { it.firstCharToUpperCase() }.firstCharToLowerCase()
-		override fun joinToString(value: Iterable<String>) = value.joinToString("") { it.firstCharToUpperCase() }.firstCharToLowerCase()
-		override fun joinToString(value: Sequence<String>) = value.joinToString("") { it.firstCharToUpperCase() }.firstCharToLowerCase()
+		override fun joinToString(value: Array<String>) =
+			value.joinToString("") { it.firstCharToUpperCase() }.firstCharToLowerCase()
+
+		override fun joinToString(value: Iterable<String>) =
+			value.joinToString("") { it.firstCharToUpperCase() }.firstCharToLowerCase()
+
+		override fun joinToString(value: Sequence<String>) =
+			value.joinToString("") { it.firstCharToUpperCase() }.firstCharToLowerCase()
 	}
 
 	/**
 	 * 以单词边界分隔，所有单词首字母大写。
 	 *
-	 * 示例：`PascalCase`
+	 * 示例：`PascalCaseFormat`
 	 */
-	object PascalCase : LetterCase {
+	object PascalCaseFormat : CaseFormat {
 		private val regex = """\$?(?:[A-Z][a-z]+|[A-Z]+)(?:\$?[A-Z][a-z]+\$?|\$?[A-Z]+\$?|\d+)+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.splitWords().split(' ')
@@ -233,7 +238,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`snake_case`
 	 */
-	object SnakeCase : LetterCase {
+	object SnakeCaseFormat : CaseFormat {
 		private val regex = """\$?[a-z]+(?:_(?:\$?[a-z]+\$?|\d+))+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.split('_')
@@ -248,7 +253,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`SCREAMING_SNAKE_CASE`
 	 */
-	object ScreamingSnakeCase : LetterCase {
+	object ScreamingSnakeCaseFormat : CaseFormat {
 		private val regex = """\$?[A-Z]+(?:_(?:\$?[A-Z]+|\d+))+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.split('_')
@@ -263,7 +268,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`Underscore_words`
 	 */
-	object UnderscoreWords : LetterCase {
+	object UnderscoreWords : CaseFormat {
 		private val regex = """_*[a-zA-Z$]+(?:_+(?:[a-zA-Z$]+|\d+))+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.split('_')
@@ -278,7 +283,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`kebab-case`
 	 */
-	object KebabCase : LetterCase {
+	object KebabCaseFormat : CaseFormat {
 		private val regex = """[a-z]+(?:-(?:[a-z]+|\d+))+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.split('-')
@@ -293,7 +298,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`KEBAB-UPPER-CASE`
 	 */
-	object KebabUpperCase : LetterCase {
+	object KebabUpperCaseFormat : CaseFormat {
 		private val regex = """[A-Z]+(?:-(?:[A-Z]+|\d+))+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.split('-')
@@ -308,7 +313,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`Hyphen-words`
 	 */
-	object HyphenWords : LetterCase {
+	object HyphenWords : CaseFormat {
 		private val regex = """-*[a-zA-Z]+(?:-+(?:[a-zA-Z]+|\d+))+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.split('-')
@@ -322,14 +327,14 @@ interface LetterCase:Component {
 	/**
 	 * 类路径的字母格式。
 	 */
-	interface PathLikeLetterCase
+	interface PathLikeCaseFormat
 
 	/**
 	 * 以单个点分隔的路径。
 	 *
 	 * 示例：`doc.path`
 	 */
-	object ReferencePath : LetterCase,PathLikeLetterCase {
+	object ReferencePath : CaseFormat, PathLikeCaseFormat {
 		private val regex = """[a-zA-Z_\-$]+(?:\.[a-zA-Z_\-$]+)+""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.split('.')
@@ -344,7 +349,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`linux/path`
 	 */
-	object LinuxPath : LetterCase,PathLikeLetterCase {
+	object LinuxPath : CaseFormat, PathLikeCaseFormat {
 		private val regex = """/?[^/\\\s]+(?:/[^/\\\s]+]+)+/?""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.trim('/').split('/')
@@ -359,7 +364,7 @@ interface LetterCase:Component {
 	 *
 	 * 示例：`windows\path`
 	 */
-	object WindowsPath : LetterCase,PathLikeLetterCase {
+	object WindowsPath : CaseFormat, PathLikeCaseFormat {
 		private val regex = """\\?[^/\\\s]+(?:\\[^/\\\s]+]+)+\\?""".toRegex()
 		override fun matches(value: String): Boolean = regex.matches(value)
 		override fun split(value: String) = value.trim('\\').split('\\')