+++ breeze-core/src/test/kotlin/extension/BreezeComponentExtensionsTest.kt
@@ -42,15 +42,15 @@ class BreezeComponentExtensionsTest {
 
 	@Test
 	fun caseTypeTest() {
-		assertEquals("abcAbcAbc", "AbcAbcAbc".switchCaseBy(LetterCase.CamelCase))
-		assertEquals("AbcAbcAbc", "abcAbcAbc".switchCaseBy(LetterCase.PascalCase))
-		assertEquals("abc-abc-abc", "ABC_ABC_ABC".switchCaseBy(LetterCase.KebabCase))
-		assertEquals("ABC_ABC_ABC", "abc-abc-abc".switchCaseBy(LetterCase.ScreamingSnakeCase))
-		assertEquals("abc abc", "AbcAbc".switchCaseBy(LetterCase.LowerCaseWords))
-		assertEquals("Abc Abc", "abcAbc".switchCaseBy(LetterCase.CapitalizedWords))
-		assertEquals("abcAbc", "Abc Abc".switchCaseBy(LetterCase.CamelCase))
-		assertEquals("abcabc", "AbcAbc".switchCaseBy(LetterCase.LowerCase))
-		assertEquals("ABCABC", "ABcABc".switchCaseBy(LetterCase.UpperCase))
+		assertEquals("abcAbcAbc", "AbcAbcAbc".switchCaseBy(CaseFormat.CamelCaseFormat))
+		assertEquals("AbcAbcAbc", "abcAbcAbc".switchCaseBy(CaseFormat.PascalCaseFormat))
+		assertEquals("abc-abc-abc", "ABC_ABC_ABC".switchCaseBy(CaseFormat.KebabCaseFormat))
+		assertEquals("ABC_ABC_ABC", "abc-abc-abc".switchCaseBy(CaseFormat.ScreamingSnakeCaseFormat))
+		assertEquals("abc abc", "AbcAbc".switchCaseBy(CaseFormat.LowerCaseFormatWords))
+		assertEquals("Abc Abc", "abcAbc".switchCaseBy(CaseFormat.CapitalizedWords))
+		assertEquals("abcAbc", "Abc Abc".switchCaseBy(CaseFormat.CamelCaseFormat))
+		assertEquals("abcabc", "AbcAbc".switchCaseBy(CaseFormat.LowerCaseFormat))
+		assertEquals("ABCABC", "ABcABc".switchCaseBy(CaseFormat.UpperCaseFormat))
 	}
 
 	@Test
@@ -65,40 +65,40 @@ class BreezeComponentExtensionsTest {
 		assertFalse { "/foo/bar".matchesBy("/foo/bar/123") }
 		assertFalse { "/foo/bar".matchesBy("/{a}/bar/123") }
 
-		assertTrue { "".matchesBy("", PathPattern.AntPath) }
-		assertTrue { "/foo/bar".matchesBy("/foo/bar", PathPattern.AntPath) }
-		assertTrue { "/foo/bar".matchesBy("/foo/b?r", PathPattern.AntPath) }
-		assertTrue { "/foo/bar".matchesBy("/foo/b*r", PathPattern.AntPath) }
-		assertTrue { "/foo/bar".matchesBy("/foo/b*", PathPattern.AntPath) }
-		assertTrue { "/foo/bar".matchesBy("/foo/bar*", PathPattern.AntPath) }
-		assertTrue { "/foo/bar".matchesBy("/foo/**", PathPattern.AntPath) }
-		assertTrue { "/foo/bar".matchesBy("/foo/bar**", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa".matchesBy("/f**o/aaa", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa".matchesBy("/f*/aaa", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa".matchesBy("/f**/aaa", PathPattern.AntPath) }
-		assertTrue { "/foo/bar/var".matchesBy("/foo/bar/**", PathPattern.AntPath) }
-		assertTrue { "/foo/bar/var".matchesBy("/foo/*/var", PathPattern.AntPath) }
-		assertTrue { "/foo/bar/var".matchesBy("/foo/**/var", PathPattern.AntPath) }
-		assertTrue { "/foo/bar/var".matchesBy("/foo/*r/var", PathPattern.AntPath) }
-		assertFalse { "/foo/bar/var".matchesBy("/foo/*/va", PathPattern.AntPath) }
-		assertFalse { "/foo/bar/var".matchesBy("/foo/**/va", PathPattern.AntPath) }
-		assertFalse { "/foo/bar/var".matchesBy("/foo/*r/va", PathPattern.AntPath) }
-		assertTrue { "/foo/bar/var".matchesBy("/foo/bar/{var}", PathPattern.AntPath) }
-		assertTrue { "/foo/bar/var".matchesBy("/foo/{bar}/var", PathPattern.AntPath) }
-		assertTrue { "".matchesBy("", PathPattern.AntPath) }
-		assertTrue { "/foo".matchesBy("/foo", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa".matchesBy("/foo/{a}", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa".matchesBy("/f?o/{a}", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa".matchesBy("/f*o/{a}", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa".matchesBy("/f**o/{a}", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa".matchesBy("/f*/{a}", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa".matchesBy("/f**/{a}", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa/bar/bbb".matchesBy("/foo/{a}/bar/{b}", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa/bar/bbb".matchesBy("/foo/{a}/*/{b}", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa/bar/bbb".matchesBy("/foo/{a}/**/{b}", PathPattern.AntPath) }
-		assertTrue { "/foo/aaa/bar/bbb".matchesBy("/foo/{a}/*ar/{b}", PathPattern.AntPath) }
+		assertTrue { "".matchesBy("", PathFormat.AntPath) }
+		assertTrue { "/foo/bar".matchesBy("/foo/bar", PathFormat.AntPath) }
+		assertTrue { "/foo/bar".matchesBy("/foo/b?r", PathFormat.AntPath) }
+		assertTrue { "/foo/bar".matchesBy("/foo/b*r", PathFormat.AntPath) }
+		assertTrue { "/foo/bar".matchesBy("/foo/b*", PathFormat.AntPath) }
+		assertTrue { "/foo/bar".matchesBy("/foo/bar*", PathFormat.AntPath) }
+		assertTrue { "/foo/bar".matchesBy("/foo/**", PathFormat.AntPath) }
+		assertTrue { "/foo/bar".matchesBy("/foo/bar**", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa".matchesBy("/f**o/aaa", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa".matchesBy("/f*/aaa", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa".matchesBy("/f**/aaa", PathFormat.AntPath) }
+		assertTrue { "/foo/bar/var".matchesBy("/foo/bar/**", PathFormat.AntPath) }
+		assertTrue { "/foo/bar/var".matchesBy("/foo/*/var", PathFormat.AntPath) }
+		assertTrue { "/foo/bar/var".matchesBy("/foo/**/var", PathFormat.AntPath) }
+		assertTrue { "/foo/bar/var".matchesBy("/foo/*r/var", PathFormat.AntPath) }
+		assertFalse { "/foo/bar/var".matchesBy("/foo/*/va", PathFormat.AntPath) }
+		assertFalse { "/foo/bar/var".matchesBy("/foo/**/va", PathFormat.AntPath) }
+		assertFalse { "/foo/bar/var".matchesBy("/foo/*r/va", PathFormat.AntPath) }
+		assertTrue { "/foo/bar/var".matchesBy("/foo/bar/{var}", PathFormat.AntPath) }
+		assertTrue { "/foo/bar/var".matchesBy("/foo/{bar}/var", PathFormat.AntPath) }
+		assertTrue { "".matchesBy("", PathFormat.AntPath) }
+		assertTrue { "/foo".matchesBy("/foo", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa".matchesBy("/foo/{a}", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa".matchesBy("/f?o/{a}", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa".matchesBy("/f*o/{a}", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa".matchesBy("/f**o/{a}", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa".matchesBy("/f*/{a}", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa".matchesBy("/f**/{a}", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa/bar/bbb".matchesBy("/foo/{a}/bar/{b}", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa/bar/bbb".matchesBy("/foo/{a}/*/{b}", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa/bar/bbb".matchesBy("/foo/{a}/**/{b}", PathFormat.AntPath) }
+		assertTrue { "/foo/aaa/bar/bbb".matchesBy("/foo/{a}/*ar/{b}", PathFormat.AntPath) }
 
-		"/foo/bar/bar/bar".matchesBy("/foo/*/b?r/**", PathPattern.AntPath)
+		"/foo/bar/bar/bar".matchesBy("/foo/*/b?r/**", PathFormat.AntPath)
 	}
 
 	@Test
@@ -109,20 +109,20 @@ class BreezeComponentExtensionsTest {
 		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/bar/{b}"))
 		assertEquals(mapOf(), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/bar/{b}/far"))
 
-		assertEquals(mapOf(), "".resolveVariablesBy("", PathPattern.AntPath))
-		assertEquals(mapOf(), "/foo".resolveVariablesBy("/foo", PathPattern.AntPath))
-		assertEquals(mapOf("a" to "aaa"), "/foo/aaa".resolveVariablesBy("/foo/{a}", PathPattern.AntPath))
-		assertEquals(mapOf("a" to "aaa"), "/foo/aaa".resolveVariablesBy("/f?o/{a}", PathPattern.AntPath))
-		assertEquals(mapOf("a" to "aaa"), "/foo/aaa".resolveVariablesBy("/f*o/{a}", PathPattern.AntPath))
-		assertEquals(mapOf("a" to "aaa"), "/foo/aaa".resolveVariablesBy("/f**o/{a}", PathPattern.AntPath))
-		assertEquals(mapOf("a" to "aaa"), "/foo/aaa".resolveVariablesBy("/f*/{a}", PathPattern.AntPath))
-		assertEquals(mapOf("a" to "aaa"), "/foo/aaa".resolveVariablesBy("/f**/{a}", PathPattern.AntPath))
-		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/bar/{b}", PathPattern.AntPath))
-		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/*/{b}", PathPattern.AntPath))
-		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/**/{b}", PathPattern.AntPath))
-		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/*ar/{b}", PathPattern.AntPath))
-		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/?ar/{b}", PathPattern.AntPath))
-		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/**ar/{b}", PathPattern.AntPath))
+		assertEquals(mapOf(), "".resolveVariablesBy("", PathFormat.AntPath))
+		assertEquals(mapOf(), "/foo".resolveVariablesBy("/foo", PathFormat.AntPath))
+		assertEquals(mapOf("a" to "aaa"), "/foo/aaa".resolveVariablesBy("/foo/{a}", PathFormat.AntPath))
+		assertEquals(mapOf("a" to "aaa"), "/foo/aaa".resolveVariablesBy("/f?o/{a}", PathFormat.AntPath))
+		assertEquals(mapOf("a" to "aaa"), "/foo/aaa".resolveVariablesBy("/f*o/{a}", PathFormat.AntPath))
+		assertEquals(mapOf("a" to "aaa"), "/foo/aaa".resolveVariablesBy("/f**o/{a}", PathFormat.AntPath))
+		assertEquals(mapOf("a" to "aaa"), "/foo/aaa".resolveVariablesBy("/f*/{a}", PathFormat.AntPath))
+		assertEquals(mapOf("a" to "aaa"), "/foo/aaa".resolveVariablesBy("/f**/{a}", PathFormat.AntPath))
+		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/bar/{b}", PathFormat.AntPath))
+		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/*/{b}", PathFormat.AntPath))
+		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/**/{b}", PathFormat.AntPath))
+		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/*ar/{b}", PathFormat.AntPath))
+		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/?ar/{b}", PathFormat.AntPath))
+		assertEquals(mapOf("a" to "aaa", "b" to "bbb"), "/foo/aaa/bar/bbb".resolveVariablesBy("/foo/{a}/**ar/{b}", PathFormat.AntPath))
 	}
 
 	@Test
@@ -141,15 +141,15 @@ class BreezeComponentExtensionsTest {
 		assertEquals(list, list.queryBy("/{name}"))
 		assertEquals(listOf(2, 3, 4), list.queryBy("/1/{name}"))
 
-		assertEquals(listOf(list), list.queryBy("", PathPattern.ReferencePath))
-		assertEquals(listOf(1), list.queryBy("[0]", PathPattern.ReferencePath))
-		assertEquals(listOf(3), list.queryBy("[1][1]", PathPattern.ReferencePath))
-		assertEquals(list, list.getBy("", PathPattern.ReferencePath))
-		assertEquals(1, list.getBy("[0]", PathPattern.ReferencePath))
-		assertEquals(3, list.getBy("[1][1]", PathPattern.ReferencePath))
-		assertEquals<Any?>(1, list.getOrNullBy("[0]", PathPattern.ReferencePath))
-		assertEquals(1, list.getOrElseBy("[0]", PathPattern.ReferencePath) { 111 })
-		assertEquals(null, list.getOrNullBy("[111]", PathPattern.ReferencePath))
-		assertEquals(111, list.getOrElseBy("[111]", PathPattern.ReferencePath) { 111 })
+		assertEquals(listOf(list), list.queryBy("", PathFormat.ReferencePath))
+		assertEquals(listOf(1), list.queryBy("[0]", PathFormat.ReferencePath))
+		assertEquals(listOf(3), list.queryBy("[1][1]", PathFormat.ReferencePath))
+		assertEquals(list, list.getBy("", PathFormat.ReferencePath))
+		assertEquals(1, list.getBy("[0]", PathFormat.ReferencePath))
+		assertEquals(3, list.getBy("[1][1]", PathFormat.ReferencePath))
+		assertEquals<Any?>(1, list.getOrNullBy("[0]", PathFormat.ReferencePath))
+		assertEquals(1, list.getOrElseBy("[0]", PathFormat.ReferencePath) { 111 })
+		assertEquals(null, list.getOrNullBy("[111]", PathFormat.ReferencePath))
+		assertEquals(111, list.getOrElseBy("[111]", PathFormat.ReferencePath) { 111 })
 	}
 }